Styled and get FS
Solved the issue with styled components not rendering from the UI library.

We did the tests on a test folder called newui
I am uploading also the test folder in this commit, i will delete it in the next.
In this way we can have a reference to the package which was working.